refactor : 사원 목록 페이지 관심사 분리
